[ChamSys Ltd] MQ50/70HD SD card: Disable SDR104 mode, enable high-speed mode
UHS-1 modes are unreliable on the MQ50/70 design. Perhaps the I/O lines are not switching to 1.8V, or
the base PCB cannot cope with 1.8V signals or the higher speeds.

Samsung Pro Endurance works most of the time, but a small number of consoles report I/O errors in UHS-1 modes.

SanDisk and Integral always produce I/O errors in UHS-1 modes.